requirements.txt
Flask==2.1.2
requests==2.28.2
python-dotenv==1.0.0

flask
Flask-SQLAlchemy
flask-login
